feat: add polling for wise connector
