+++ components/payments/internal/app/connectors/wise/client/client.go
@@ -1,0 +1,42 @@
+package client
+
+import (
+	"fmt"
+	"net/http"
+
+	"go.opentelemetry.io/contrib/instrumentation/net/http/otelhttp"
+)
+
+const apiEndpoint = "https://api.wise.com"
+
+type apiTransport struct {
+	APIKey     string
+	underlying http.RoundTripper
+}
+
+func (t *apiTransport) RoundTrip(req *http.Request) (*http.Response, error) {
+	req.Header.Add("Authorization", fmt.Sprintf("Bearer %s", t.APIKey))
+
+	return t.underlying.RoundTrip(req)
+}
+
+type Client struct {
+	httpClient *http.Client
+}
+
+func (w *Client) endpoint(path string) string {
+	return fmt.Sprintf("%s/%s", apiEndpoint, path)
+}
+
+func NewClient(apiKey string) *Client {
+	httpClient := &http.Client{
+		Transport: &apiTransport{
+			APIKey:     apiKey,
+			underlying: otelhttp.NewTransport(http.DefaultTransport),
+		},
+	}
+
+	return &Client{
+		httpClient: httpClient,
+	}
+}
