+++ components/payments/internal/app/connectors/wise/config.go
@@ -3,11 +3,13 @@ package wise
 import (
 	"encoding/json"
 
+	"github.com/formancehq/payments/internal/app/connectors"
 	"github.com/formancehq/payments/internal/app/connectors/configtemplate"
 )
 
 type Config struct {
-	APIKey string `json:"apiKey" yaml:"apiKey" bson:"apiKey"`
+	APIKey        string              `json:"apiKey" yaml:"apiKey" bson:"apiKey"`
+	PollingPeriod connectors.Duration `json:"pollingPeriod" yaml:"pollingPeriod" bson:"pollingPeriod"`
 }
 
 // String obfuscates sensitive fields and returns a string representation of the config.
@@ -32,6 +34,7 @@ func (c Config) BuildTemplate() (string, configtemplate.Config) {
 	cfg := configtemplate.NewConfig()
 
 	cfg.AddParameter("apiKey", configtemplate.TypeString, true)
+	cfg.AddParameter("pollingPeriod", configtemplate.TypeDurationNs, false)
 
 	return Name.String(), cfg
 }
