+++ components/payments/internal/app/connectors/wise/task_fetch_profiles.go
@@ -4,19 +4,18 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/formancehq/payments/internal/app/connectors/wise/client"
 	"github.com/formancehq/payments/internal/app/models"
-
 	"github.com/formancehq/payments/internal/app/task"
-
 	"github.com/formancehq/stack/libs/go-libs/logging"
 )
 
-func taskFetchProfiles(logger logging.Logger, client *client) task.Task {
+func taskFetchProfiles(logger logging.Logger, client *client.Client) task.Task {
 	return func(
 		ctx context.Context,
 		scheduler task.Scheduler,
 	) error {
-		profiles, err := client.getProfiles()
+		profiles, err := client.GetProfiles()
 		if err != nil {
 			return err
 		}
